error handling for users

+++ backend/controllers/user.js
@@ -22,10 +22,10 @@ exports.signup = (req, res, next) => {
             User.create({
                 ...newUser
             })
-                .then(user => res.status(200).json(user))
-                .catch(error => res.status(400).json({ error }));
+                .then(user => res.status(201).json(user))
+                .catch(error => res.status(400).json({error}));
         })
-        .catch(error => res.status(400).json({ error }));
+        .catch(error => res.status(500).json({error}));
 }
 
 exports.login = (req, res, next) => {
@@ -34,7 +34,7 @@ exports.login = (req, res, next) => {
             bcrypt.compare(req.body.password, user.password)
                 .then(valid => {
                     if (!valid) {
-                        return res.status(401).json({ error: 'Mot de passe incorrect' });
+                        return res.status(401).json({error: 'Utilisateur ou mot de passe incorrect'});
                     }
                     res.status(200).json({
                         userId: user.id,
@@ -45,9 +45,9 @@ exports.login = (req, res, next) => {
                         )
                     });
                 })
-                .catch(error => res.status(500).json({ error }));
+                .catch(error => res.status(500).json({error}));
         })
-        .catch(error => res.status(401).json({ error: "Utilisateur non trouvé" }));
+        .catch(() => res.status(401).json({error: 'Utilisateur ou mot de passe incorrect'}));
 };
 
 exports.modifyUser = (req, res, next) => {
@@ -63,19 +63,23 @@ exports.modifyUser = (req, res, next) => {
 
     const updateUser = () => {
         User.update(userUpdated, { where: { id: req.params.id } })
-            .then(() => res.status(200).json({ message: "Utilisateur modifié" }))
-            .catch(error => res.status(400).json({ error }));
+            .then(() => res.status(201).json({ message: 'Utilisateur modifié' }))
+            .catch(error => res.status(400).json({error}));
     }
 
     if (req.file) {
         User.findOne({ where: { id: req.params.id } })
             .then(user => {
-                const filename = user.profilePicture.split('/images')[1];
-                fs.unlink(`images/${filename}`, () => {
+                if(user.profilePicture) {
+                    const filename = user.profilePicture.split('/images')[1];
+                    fs.unlink(`images/${filename}`, () => {
+                        updateUser();
+                    })
+                }else{
                     updateUser();
-                })
+                }
             })
-            .catch(error => res.status(500).json({ error }));
+            .catch(error => res.status(404).json({error}));
     } else {
         updateUser();
     }
@@ -89,13 +93,13 @@ exports.deleteUser = (req, res, next) => {
             fs.unlink(`images/${filename}`, () => {
                 User.destroy({ where: { id: req.params.id } })
                     .then(() => res.status(200).json({ message: 'Utilisateur supprimé' }))
-                    .catch(error => res.status(400).json({ error }));
+                    .catch(error => res.status(400).json({error}));
             })
         })
-        .catch(error => res.status(500).json({ error }));
+        .catch(error => res.status(404).json({error}));
     } else {
         User.destroy({ where: { id: req.params.id } })
                     .then(() => res.status(200).json({ message: 'Utilisateur supprimé' }))
-                    .catch(error => res.status(400).json({ error }));
+                    .catch(error => res.status(400).json({error}));
     }  
 }
